chore: refactor CadNode and related into proper package (#2173)
* chore: refactor CadNode and related into proper package

* chore: refactor batching components out of rendering

* chore: move files into proper folders

* chore: remove unused package dependencies

* chore: remove unused package imports and restore tests

* chore: remove dead code

* chore: add mock for materials
