+++ viewer/core/src/public/types.ts
@@ -6,7 +6,7 @@ import * as THREE from 'three';
 
 import { LoadingState } from '@reveal/model-base';
 import { RenderOptions } from '@reveal/rendering';
-import { InternalRevealCadOptions } from '@reveal/cad-model';
+import { InternalRevealCadOptions } from '@reveal/cad-geometry-loaders';
 
 /**
  * @property logMetrics Might be used to disable usage statistics.
