+++ viewer/packages/cad-geometry-loaders/index.ts
@@ -9,3 +9,6 @@ export { CadModelUpdateHandler } from './src/CadModelUpdateHandler';
 export { SectorCuller } from './src/sector/culling/SectorCuller';
 
 export { createV8SectorCuller } from './src/sector/culling/createV8SectorCuller';
+export { CadManager } from './src/CadManager';
+export { createCadManager } from './src/createCadManager';
+export { InternalRevealCadOptions } from './src/InternalRevealCadOptions';
