+++ viewer/packages/cad-geometry-loaders/src/InternalRevealCadOptions.ts
