+++ viewer/packages/cad-geometry-loaders/src/sector/culling/types.ts
@@ -8,7 +8,7 @@ import { CadModelMetadata, LevelOfDetail, WantedSector } from '@reveal/cad-parse
 
 import { CadLoadingHints } from '../../CadLoadingHints';
 import { CadModelBudget } from '../../CadModelBudget';
-import { CadNode } from '@reveal/rendering';
+import { CadNode } from '@reveal/cad-model';
 
 export interface DetermineSectorsInput {
   camera: THREE.PerspectiveCamera;
