+++ viewer/packages/cad-model/package.json
@@ -8,7 +8,6 @@
     "lint": "yarn ws:lint"
   },
   "dependencies": {
-    "@reveal/cad-geometry-loaders": "workspace:*",
     "@reveal/cad-parsers": "workspace:*",
     "@reveal/cad-styling": "workspace:*",
     "@reveal/metrics": "workspace:*",
@@ -17,6 +16,7 @@
     "@reveal/nodes-api": "workspace:*",
     "@reveal/rendering": "workspace:*",
     "@reveal/sector-loader": "workspace:*",
+    "@reveal/sector-parser": "workspace:*",
     "@reveal/utilities": "workspace:*"
   }
 }
