+++ viewer/packages/cad-model/src/picking/PickingHandler.ts
@@ -7,12 +7,12 @@ import { IntersectInput } from '@reveal/model-base';
 import {
   BasicPipelineExecutor,
   CadMaterialManager,
-  CadNode,
   CadGeometryRenderModePipelineProvider,
   RenderMode,
   RenderPipelineProvider
 } from '@reveal/rendering';
 import { SceneHandler, WebGLRendererStateHelper } from '@reveal/utilities';
+import { CadNode } from '../wrappers/CadNode';
 
 type PickingInput = {
   normalizedCoords: {
