+++ viewer/packages/cad-model/src/utilities/BoundingBoxClipper.test.ts
