+++ viewer/packages/cad-model/src/utilities/BoundingBoxClipper.ts
