+++ viewer/packages/cad-model/src/utilities/NodeIdAndTreeIndexMaps.ts
