+++ viewer/packages/cad-model/src/utilities/callActionWithIndicesAsync.ts
