+++ viewer/packages/rendering/package.json
@@ -11,8 +11,6 @@
     "@reveal/cad-parsers": "workspace:*",
     "@reveal/cad-styling": "workspace:*",
     "@reveal/logger": "workspace:*",
-    "@reveal/sector-loader": "workspace:*",
-    "@reveal/sector-parser": "workspace:*",
     "@reveal/utilities": "workspace:*"
   },
   "sideEffects": false,
