+++ viewer/test-utilities/src/createCadModel.ts
@@ -2,9 +2,9 @@
  * Copyright 2021 Cognite AS
  */
 import { V8SectorRepository } from '../../packages/sector-loader';
-import { Cognite3DModel } from '../../packages/cad-model';
+import { CadNode, Cognite3DModel } from '../../packages/cad-model';
 import { NodesLocalClient } from '../../packages/nodes-api';
-import { CadMaterialManager, CadNode } from '../../packages/rendering';
+import { CadMaterialManager } from '../../packages/rendering';
 
 import { createCadModelMetadata } from './createCadModelMetadata';
 import { generateV8SectorTree } from './createSectorMetadata';
